requirements.txt
langchain>=0.1.0
langchain-groq>=0.2.1
langchain-core>=0.1.8
langgraph>=0.0.12
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
slowapi>=0.1.9
redis>=5.0.1
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
requests>=2.31.0
